Fix deserialization error for some survey configurations
The SDK may fail to deserialize the response from the server when
requesting the survey configuration. When this happens the error message,
"Could not decode eligibility configuration request", appears in the
console and the survey will not be shown.

The client-side eligibility check performed by the SDK uses the survey
configuration which may have null values for the `initialSurveyDelay` and
`recurringPeriod` fields. Since the SDK expects these to always be
non-nil, the response fails to be decoded leading to the error. The SDK
should really treat these fields as optional since they don't apply in
some configurations. A bonus to changing this is that it allows for a
more specific eligibility error message when the SDK is set up as a
one-time survey.